adding new testCommand package into contentGathering service
This enables the testing of commands on a target endpoint. The commands are transformed/updated as the framework would do normally in another job, but this allows tactical testing of environmental differences, like elevated account access and command output variance across target endpoints... without continuing to rerun a standard job.