query creation on filesystem
Will probably take all query definition writing to the filesystem out in future release, but for now - ensure the directory exists before writing.  If it was removed for cleanup purposes, this would previously cause a false negative. The definition would be saved to the DB, but error out on the file system replica.