find jobs to create NameRecords
The best approach for DNS context is to get zone transfers or WMI dumps. But for an alternative, since names are often times required for other discovery or reporting tasks, this enhances the find SSH and PowerShell jobs to do a local name (on remote server endpoints) lookup of IPs found.