constructAndSendData to be synchronous operation
Main change here is to make service-to-client communication happen synchronously, so that multiple jobs running at the same time (called in non-reactor threads) will not undesirably merge partial data on the sendLine operation. This is rare but was noticed, hence the fix.  Note that due to the issues with epoll on Linux, the constructAndSendData function has to check if the reactor is loaded every time (via import statement)... which hopefully can be pulled out of here later.